refactor(projects): address a project by query param instead of a path segment
/projects/[id] could not work on a static export. The export emits one document
per id returned by generateStaticParams, and project ids are database rows, so
every real id resolved to no document: CloudFront fell back to another page's
shell and the client router rendered not-found. Every project deep link 404'd,
including the ones the new navbar dropdown produces.

/projects?id=1 is a single prerendered document that reads the id at runtime, so
deep links, refreshes and in-app navigation all work with no CloudFront rules
and no placeholder shell to keep in step with the infrastructure.

The navbar now takes the active project as a prop rather than parsing it out of
the path; reading the query param there would force a Suspense boundary onto
every page that renders the rail.
