+++ apps/frontend/src/app/page.tsx
@@ -20,12 +20,13 @@ import FullPageSpinner from './components/FullPageSpinner';
  *
  * It also absorbs the CloudFront SPA fallback: the distribution rewrites every
  * 403/404 to `/index.html` (see infrastructure/aws/frontend_hosting.tf), so a
- * deep link to a path with no exported document — `/projects/7/`, say — is served
- * *this* document. We detect that and hand the URL back to the client router.
+ * deep link to a path with no exported document is served *this* document. We
+ * detect that and hand the URL back to the client router.
  *
- * Follow-up for whoever owns the infra: adding a CloudFront Function rule that
- * maps `/projects/*` to the dynamic route's document would let those deep links
- * hydrate directly instead of round-tripping through here.
+ * Note that landing here is a last resort — the client router can only render
+ * routes the export actually emitted, so a path with no document of its own
+ * still ends at not-found. That is why per-record pages take the id as a query
+ * param (`/projects?id=1`) instead of a path segment: the document exists.
  */
 
 const SPA_FALLBACK_KEY = 'branch_spa_fallback_path';
